changed name of the project from 'experience' to 'character'.
