+++ src/main/java/com/knightcharacter/app/rest/response/TodoResponse.java
@@ -1,0 +1,5 @@
+package com.knightcharacter.app.rest.response;
+
+public class TodoResponse {
+
+}
