+++ src/main/java/com/knightcharacter/app/service/ExperienceService.java
@@ -1,0 +1,8 @@
+package com.knightcharacter.app.service;
+
+import com.knightcharacter.app.domain.TodoVO;
+
+public interface ExperienceService {
+
+    void calculateExperience(TodoVO todoVO);
+}
