--- src/main/java/com/knightexperience/app/rest/response/TodoResponse.java
@@ -1,5 +1,0 @@
-package com.knightexperience.app.rest.response;
-
-public class TodoResponse {
-
-}
