--- src/main/java/com/knightexperience/app/service/ExperienceService.java
@@ -1,8 +1,0 @@
-package com.knightexperience.app.service;
-
-import com.knightexperience.app.domain.TodoVO;
-
-public interface ExperienceService {
-
-    void calculateExperience(TodoVO todoVO);
-}
