+++ src/test/java/com/knightcharacter/app/AppApplicationTests.java
@@ -1,4 +1,4 @@
-package com.knightexperience.app;
+package com.knightcharacter.app;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
